feat(logger): export RotFileConfig so the rotfile writer is configurable from pkg/v1 (#93)
The "rotfile" writer's config struct lived in internal/service/writer/rotfile,
so pkg/v1/logger could not re-export it the way it does ConsoleConfig and
FileConfig — a consumer could name the "rotfile" writer in a WriterSpec but had
no public type to build its Config without reaching into internal/* (forbidden
by the internal/ firewall).

Move the canonical struct to internal/core/writer (RotFileConfig), mirroring
ConsoleConfig / FileConfig, and alias the in-package service type
(rotfile.Config = corewriter.RotFileConfig) so type identity is unchanged —
Open's type-assertion and every existing caller compile untouched. pkg/v1/logger
now re-exports it as logger.RotFileConfig, closing the gap.
